requirements.txt
crewai>=0.11.0
crewai-tools>=0.0.1
python-dotenv>=1.0.0 
openai>=1.75.0